major release: mail filters, search refactor, browser push, and more

+++ skiff-mail-web/components/Settings/Account/DeleteAccount/DeleteAccount.tsx
@@ -1,24 +1,23 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { TitleActionSection } from 'skiff-front-utils';
+import { ConfirmDeleteAccountModal, TitleActionSection, useRequiredCurrentUserData } from 'skiff-front-utils';
 import { LoginSrpRequest } from 'skiff-graphql';
 
-import { useRequiredCurrentUserData } from '../../../../apollo/currentUser';
 import { useMailLogout } from '../../../../hooks/useMailLogout';
 import { skemailModalReducer } from '../../../../redux/reducers/modalReducer';
 import { deleteAccount } from '../../../../utils/userUtils';
 import ConfirmPasswordDialog from '../../../shared/ConfirmPasswordDialog';
 
-const STEPS_ACCOUNT_DELETE = {
-  INSTRUCTION_PAGE: 0,
-  LOADING: 1,
-  CONFIRM_CHANGE: 2
-};
+enum DeleteModalType {
+  NONE,
+  CONFIRM_PASSWORD,
+  CONFIRM_DELETION
+}
 
 /**
- * Component for rendering the change password section in AccountSettings.
+ * Component for rendering the interface to delete one's own account.
  */
-function DeleteAccountDialog() {
+function DeleteAccount() {
   const userData = useRequiredCurrentUserData();
   const dispatch = useDispatch();
   const logout = useMailLogout();
@@ -27,54 +26,59 @@ function DeleteAccountDialog() {
     dispatch(skemailModalReducer.actions.setOpenModal(undefined));
   }, [dispatch]);
 
-  // Step initially set to INSTRUCTION_PAGE when dialog loads.
-  // Later set to LOADING or CONFIRM_CHANGE as deletion progresses.
-  const [stepAccountDelete, setStepAccountDelete] = useState(STEPS_ACCOUNT_DELETE.INSTRUCTION_PAGE);
-  const [showDeleteAccountDialog, setShowDeleteAccountDialog] = useState(false);
+  const [currentModal, setCurrentModal] = useState(DeleteModalType.NONE);
+  const [loginSrpRequest, setLoginSrpRequest] = useState<LoginSrpRequest | null>(null);
 
-  async function performAccountDeletion(loginSrpRequest: LoginSrpRequest) {
+  const closeDeleteModal = () => setCurrentModal(DeleteModalType.NONE);
+
+  async function performAccountDeletion() {
+    if (!loginSrpRequest) {
+      console.error('Re-authentication failed -- loginSrpRequest not set correctly');
+      setCurrentModal(DeleteModalType.CONFIRM_PASSWORD);
+      return false;
+    }
     const success = await deleteAccount(userData, loginSrpRequest);
     if (success) {
-      setStepAccountDelete(STEPS_ACCOUNT_DELETE.CONFIRM_CHANGE);
-    } else {
-      setStepAccountDelete(STEPS_ACCOUNT_DELETE.INSTRUCTION_PAGE);
+      closeDeleteModal();
+      closeSettingsModal();
+      void logout();
+      return true;
     }
-    return success;
+    return false;
   }
 
-  /** Resolves async race condition with React hooks */
-  useEffect(() => {
-    if (stepAccountDelete === STEPS_ACCOUNT_DELETE.CONFIRM_CHANGE) {
-      setShowDeleteAccountDialog(false);
-      closeSettingsModal();
-      logout();
-    }
-  }, [closeSettingsModal, logout, stepAccountDelete]);
-
-  const openDeleteDialog = () => setShowDeleteAccountDialog(true);
   return (
     <>
       <TitleActionSection
         actions={[
           {
-            onClick: openDeleteDialog,
+            onClick: () => setCurrentModal(DeleteModalType.CONFIRM_PASSWORD),
             label: 'Delete',
             type: 'button',
             destructive: true
           }
         ]}
         destructive
-        subtitle='Erase all your content and data.'
+        subtitle='Erase all your content and data'
         title='Delete account'
       />
       <ConfirmPasswordDialog
         description='Please re-authenticate to confirm deletion.'
-        handleSubmit={performAccountDeletion}
-        onClose={() => setShowDeleteAccountDialog(false)}
-        open={showDeleteAccountDialog}
+        handleSubmit={(req) => {
+          setLoginSrpRequest(req);
+          setCurrentModal(DeleteModalType.CONFIRM_DELETION);
+          return new Promise(() => true);
+        }}
+        onClose={closeDeleteModal}
+        open={currentModal === DeleteModalType.CONFIRM_PASSWORD}
+      />
+      <ConfirmDeleteAccountModal
+        onClose={closeDeleteModal}
+        onConfirm={performAccountDeletion}
+        open={currentModal === DeleteModalType.CONFIRM_DELETION}
       />
     </>
   );
 }
 
-export default DeleteAccountDialog;
+export default DeleteAccount;
